Rejiggered the calculate from ints to doubles and added Euclidean

+++ src/DijkstraHeuristic.java
@@ -1,6 +1,6 @@
 public class DijkstraHeuristic implements Heuristic {
-  public int calculate(MapObject start, MapObject end) {
-        return 0;
+  public double calculate(MapObject start, MapObject end) {
+        return 0.0;
   }
   public String getName() {
     return "Dijkstra";
